+++ auth-server/src/test/java/com/truve/platform/auth/service/controller/AuthControllerTest.java
@@ -19,12 +19,10 @@ import org.springframework.test.web.servlet.ResultActions;
 import com.truve.platform.auth.service.security.AuthCookieManager;
 import com.truve.platform.auth.service.security.config.SecurityConfig;
 import com.truve.platform.auth.service.service.AuthService;
-import com.truve.platform.auth.service.domain.dto.response.AuthResponse;
 import com.truve.platform.common.exception.CustomException;
 import com.truve.platform.common.exception.ErrorCode;
 
 import jakarta.servlet.http.HttpServletResponse;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @WebMvcTest(controllers = AuthController.class)
 @Import(SecurityConfig.class)
@@ -199,37 +197,4 @@ class AuthControllerTest {
 		verify(authCookieManager).setRefreshToken(any(HttpServletResponse.class), eq("new-refresh-token"), eq(1209600L));
 	}
 
-	@Test
-	@DisplayName("내 정보 조회에 성공하면 200 OK와 사용자 정보를 반환한다.")
-	void 내정보조회_성공() throws Exception {
-		// given
-		given(authService.getMe("access-token"))
-			.willReturn(new AuthResponse.Me("user@test.com", "tester", false));
-
-		// when
-		ResultActions resultActions = mockMvc.perform(get("/api/auth/me")
-			.header("X-Token", "access-token"));
-
-		// then
-		resultActions.andExpect(status().isOk())
-			.andExpect(jsonPath("$.code").value("ok"))
-			.andExpect(jsonPath("$.data.email").value("user@test.com"))
-			.andExpect(jsonPath("$.data.nickname").value("tester"))
-			.andExpect(jsonPath("$.data.marketingInfoAgreed").value(false));
-		verify(authService).getMe("access-token");
-	}
-
-	@Test
-	@DisplayName("로그아웃에 성공하면 200 OK를 반환하고 refreshToken 쿠키를 제거한다.")
-	void 로그아웃_성공() throws Exception {
-		// when
-		ResultActions resultActions = mockMvc.perform(delete("/api/auth/logout")
-			.header("X-Token", "access-token"));
-
-		// then
-		resultActions.andExpect(status().isOk())
-			.andExpect(jsonPath("$.code").value("ok"));
-		verify(authService).logout("access-token");
-		verify(authCookieManager).clearRefreshToken(any(HttpServletResponse.class));
-	}
 }
